pyproject.toml
[project]
name = "culinary-canvas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "celery>=5.5.3",
    "django>=5.2.6",
    "django-modeltranslation>=0.19.17",
    "django-redis>=6.0.0",
    "django-rosetta>=0.10.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]
